session_supervisor: carve sub_agents.rs (pure move)
